+++ src/main/java/com/hyerijang/dailypay/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.hyerijang.dailypay.user.repository;
 
-import com.hyerijang.dailypay.user.entity.User;
+import com.hyerijang.dailypay.user.domain.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
