+++ src/main/java/com/hyerijang/dailypay/user/service/UserService.java
@@ -1,7 +1,7 @@
 package com.hyerijang.dailypay.user.service;
 
 import com.hyerijang.dailypay.auth.RegisterRequest;
-import com.hyerijang.dailypay.user.entity.User;
+import com.hyerijang.dailypay.user.domain.User;
 import com.hyerijang.dailypay.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
